Added a bunch of animations. All Scripts now Use the new Ticker

+++ Untitled Bauhaus Game/Assets/Scripts/Savegame Scripts/Save.cs
@@ -15,7 +15,7 @@ public class Save
 
 	public float CurrentMoney;
 	public int CurrentAnsehen;
-	public int CurrentPolitics;
+	public float CurrentPolitics;
 	public int StudentenAnzahl;
 	public int MaxStudenten;
